naively bump attempt of vitest in blackbox tests
